refactor: coding standards compliance - constants & modularization (v6.1.1)
## Changes

### Magic Number Elimination
- Add TIME_MS constants (SECOND, MINUTE, HOUR, DAY, WEEK) to lib/constants.ts
- Add HTTP_STATUS.GONE (410) to lib/constants/ui.ts
- Update lib/use-error-handler.ts to use TOAST_DURATION constants
- Update lib/utils.ts to use TIME_MS.DAY and TIME_MS.WEEK
- Update components/sync/oauth-buttons.tsx to use OAUTH_POPUP constants
- Update lib/sync/oauth-handshake.ts to use HTTP_STATUS.GONE

### File Size Compliance
- Reduce components/matrix-board/index.tsx from 360 to 345 lines
- Extract toDraft() helper to lib/types.ts

### OAuth Handshake Modularization
Split 389-line oauth-handshake.ts into focused modules:
- types.ts (37 lines) - Type definitions
- state.ts (74 lines) - State management
- fetcher.ts (140 lines) - API result fetching
- broadcaster.ts (104 lines) - Multi-channel broadcasting
- initializer.ts (178 lines) - Module setup & event listeners
- subscriber.ts (57 lines) - Public subscription API
- Backward-compatible re-export layer maintained

### Technical Debt
- Add ticket #90 reference to TODO in packages/mcp-server/src/tools/sync-status.ts

## Compliance
- All source files now <350 lines
- All magic numbers replaced with named constants
- 100% backward compatibility maintained
- All 21 oauth-handshake tests pass
